4-22 finish

+++ src/main/java/Question22.java
@@ -1,8 +1,8 @@
 public class Question22 implements Question {
     @Override
     public void run() {
-        long first = 0;
-        long second = 1;
+        int first = 0;
+        int second = 1;
         int N = 1000;
         System.out.print(first + " " + second + " ");
         FibonacciSequence.fibonacciRecurrenceFormula(first, second, N);
